modules/location-grpc/requirements.txt
grpcio-tools==1.54.0
kafka-python==2.0.2